# frontend/api/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    FeedStatus:
      type: string
      enum:
        - active
        - inactive
        - error
    BackendFeedStatus:
      type: string
      enum:
        - unclaimed
        - active
        - failing
        - quarantined
        - deactivated
    Tag:
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
      type: object
      additionalProperties: false
    BackendFeedStatusReason:
      type: string
      enum:
        - unknown
        - pipeline_publish_after_bookmark_failed
        - source_offline
        - source_unreachable
        - source_rate_limited
        - system_authentication_failed
        - system_configuration_invalid
        - system_source_configuration_invalid
        - system_runtime_configuration_invalid
        - system_credential_access_failed
        - system_source_payload_invalid
        - system_collector_error
        - system_pipeline_error
        - system_unexpected_error
    SourceType:
      enum:
        - bcfy_feeds
        - bcfy_calls
        - echo
        - openmhz
        - fire_notifications
      type: string
    Feed:
      properties:
        name:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        id:
          type: string
        sourceFeedId:
          type: string
        sourceUrl:
          type: string
        archiveUrl:
          type: string
        status:
          $ref: "#/components/schemas/FeedStatus"
        substatus:
          $ref: "#/components/schemas/BackendFeedStatus"
        lastHeartbeat:
          type: number
          format: double
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          type: array
        statusReason:
          $ref: "#/components/schemas/BackendFeedStatusReason"
        lastSpeechSegmentTimestamp:
          type: number
          format: double
        statusReasonDetail:
          type: string
      required:
        - name
        - sourceType
        - id
        - status
        - substatus
      type: object
      additionalProperties: false
    ListFeedsResponse:
      properties:
        feeds:
          items:
            $ref: "#/components/schemas/Feed"
          type: array
        nextToken:
          type: string
        total:
          type: number
          format: double
      required:
        - feeds
        - total
      type: object
      additionalProperties: false
...
    FeedCreate:
      properties:
        name:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        sourceFeedId:
          type: string
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          type: array
      required:
        - name
        - sourceType
        - sourceFeedId
      type: object
      additionalProperties: false
    FeedUpdate:
      properties:
        name:
          type: string
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          type: array
      required:
        - name
      type: object
      additionalProperties: false
...
  securitySchemes:
    google_id_token:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ""
          scopes: {}
      x-google-auth:
        issuer: https://accounts.google.com
        jwksUri: https://www.googleapis.com/oauth2/v3/certs
        audiences:
          - ${google_auth_client_id}
info:
  title: radio-transcription-api
  version: 1.0.0
  contact: {}
paths:
...
  /api/v1/feeds:
    get:
      operationId: ListFeeds
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ListFeedsResponse"
                  - items:
                      $ref: "#/components/schemas/Feed"
                    type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            format: int32
            type: integer
        - in: query
          name: nextToken
          required: false
          schema:
            type: string
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - in: query
          name: sourceTypes
          required: false
          schema:
            type: string
        - in: query
          name: statuses
          required: false
          schema:
            type: string
        - in: query
          name: tags
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: name
          required: false
          schema:
            type: string
      x-google-backend: radio-transcription-api
    post:
      operationId: CreateFeed
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedCreate"
      x-google-backend: radio-transcription-api
  /api/v1/feeds/{feedId}:
    get:
      operationId: GetFeed
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: path
          name: feedId
          required: true
          schema:
            type: string
      x-google-backend: radio-transcription-api
    put:
      operationId: UpdateFeed
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      description: |-
        Update an existing feed (Full override).
        The fields passed here will fully override the fields stored. There is no coalesing done, so make sure these are the final desired fields.
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: path
          name: feedId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedUpdate"
      x-google-backend: radio-transcription-api
    delete:
      operationId: DeleteFeed
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      description: |-
        Hard delete a feed (permanent delete).
        Deletes the feed, corresponding transcripts, and audio segments.
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: path
          name: feedId
          required: true
          schema:
            type: string
      x-google-backend: radio-transcription-api
  /api/v1/feeds/{feedId}/reset:
    post:
      operationId: ResetFeed
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: path
          name: feedId
          required: true
          schema:
            type: string
      x-google-backend: radio-transcription-api
  /api/v1/feeds/{feedId}/deactivate:
    post:
      operationId: DeactivateFeed
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      description: |-
        Deactivate a feed (soft delete).
        Marks the feed as deactivated to preserve historical transcripts.
      tags:
        - Feeds
      security:
        - google_id_token: []
      parameters:
        - in: path
          name: feedId
          required: true
          schema:
            type: string
      x-google-backend: radio-transcription-api
...
servers:
  - url: https://${radio_transcription_api_service_name}/
    x-google-endpoint:
      allowCors: true
...